Add Authentication service and move login function to there. Add logout function to Authentication service
