+++ pkg/services/authentication.go
@@ -1,0 +1,53 @@
+package services
+
+import (
+	"jim/twitter/pkg/dao"
+	"jim/twitter/pkg/db"
+	"jim/twitter/pkg/utils"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func Logout(c *gin.Context) {
+	accessDetails, err := utils.ExtractTokenMetadata(c.Request)
+	if err != nil {
+		c.JSON(http.StatusUnauthorized, "unauthorized")
+		return
+	}
+	deleted, delErr := utils.DeleteAuth(accessDetails.AccessUuid)
+	if delErr != nil || deleted == 0 { //if any goes wrong
+		c.JSON(http.StatusUnauthorized, "unauthorized")
+		return
+	}
+	c.JSON(http.StatusOK, "Successfully logged out")
+}
+
+func Login(c *gin.Context) {
+	var u dao.User
+	if err := c.ShouldBindJSON(&u); err != nil {
+		c.JSON(http.StatusUnprocessableEntity, "Invalid json provided")
+		return
+	}
+	user := dao.User{}
+	db.MYSQL.Find(&user)
+	//compare the user from the request, with the one we defined:
+	if user.Username != u.Username || user.Password != u.Password {
+		c.JSON(http.StatusUnauthorized, "Please provide valid login details")
+		return
+	}
+	token, err := utils.CreateToken(uint64(user.ID))
+	if err != nil {
+		c.JSON(http.StatusUnprocessableEntity, err.Error())
+		return
+	}
+	saveErr := utils.CreateAuth(uint64(user.ID), token)
+	if saveErr != nil {
+		c.JSON(http.StatusUnprocessableEntity, saveErr.Error())
+	}
+	tokens := map[string]string{
+		"access_token":  token.AccessToken,
+		"refresh_token": token.RefreshToken,
+	}
+	c.JSON(http.StatusOK, tokens)
+}
